Update install.md to include yum repo installation example
Add a roughly equivalent snippet of code for adding the openvox8-release package to some rpm based distributions.

This change omits amazonlinux and suse enterprise linux, so it's not as all in one as the debian based distro example. However it is a starting point for those people that are familiar with those platforms and any changes are welcome.